+++ NEWS.md
@@ -1,10 +1,16 @@
-# mtb 0.1.7 (commit `TBA`)
+# mtb 0.1.8 (commit `TBA`)
+
+## Major changes
+
+* bill - add comparison between 2 tables
+
+# mtb 0.1.7 (commit `cbacf062`)
 
 ## Major changes
 
 * text - add length restriction option
 * text - add functions for name, address, and slashes
-* bill - add cross comparison between multiple bill table
+* bill - add cross-count between multiple bill tables
 
 # mtb 0.1.6 (commit `d7b8929`)